CategoryService implementation

+++ admin/src/main/java/dev/coralwombat/appointment/booking/admin/rest/CategoryService.java
@@ -8,10 +8,13 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
+import dev.coralwombat.appointment.booking.dto.CategoryDTO;
 import dev.coralwombat.appointment.booking.entities.Category;
 import lombok.extern.log4j.Log4j2;
 
@@ -25,37 +28,35 @@ public class CategoryService {
 
 	@Transactional
 	@PutMapping(path = "/put")
-	public ResponseEntity<Object> put(@RequestParam(required = false) Integer id,
-			@RequestParam(required = true) String name, @RequestParam(required = false) Integer parent) {
-		log.info("CategoryService.put() called with: id=" + id + ", name=" + name + ", parent=" + parent + ".");
-		ResponseEntity<Object> response = new ResponseEntity<>(HttpStatus.OK);
+	public ResponseEntity<Object> put(@RequestBody(required = true) CategoryDTO category) {
+		log.info("CategoryService.put() called with: category=" + category.toString() + ".");
+		HttpStatus status = HttpStatus.OK;
 
-		var category = entityManager.find(Category.class, id);
-		if (category == null) {
-			category = new Category();
-			response = new ResponseEntity<>(HttpStatus.CREATED);
+		var dbCategory = entityManager.find(Category.class, category.getId());
+		if (dbCategory == null) {
+			status = HttpStatus.CREATED;
 		}
 
-		category.setId(id);
-		category.setName(name);
-		if (parent != null) {
-			var parentCategory = entityManager.find(Category.class, parent);
-			if (parentCategory == null) {
-				log.error("Could not find parent. id=" + parent);
-				return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
-			}
-			category.setParent(parentCategory);
-		}
-		entityManager.merge(category);
+		var entityCategory = new Category();
+		entityCategory.setId(category.getId());
+		entityCategory.setName(category.getName());
+		entityCategory.setParent(entityManager.find(Category.class, category.getParent()));
+		entityManager.merge(entityCategory);
 
 		log.info("CategoryService.put() finished.");
-		return response;
+		return ResponseEntity.status(status).build();
 	}
 
 	@Transactional
 	@DeleteMapping(path = "/delete")
 	public void delete(@RequestParam(required = true) Integer id) {
 		log.info("CategoryService.delete() called with: id=" + id);
+		boolean hasChildren = entityManager
+				.createQuery("SELECT c FROM Category c WHERE c.parent = :parentId", Category.class)
+				.setParameter("parentId", entityManager.find(Category.class, id)).getResultList().isEmpty();
+		if (hasChildren)
+			throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE, "Children found. id=" + id);
+
 		entityManager.createQuery("DELETE FROM Category c WHERE c.id = :id").setParameter("id", id).executeUpdate();
 		log.info("CategoryService.delete() finished.");
 	}
